+++ Communication.Test/ModBusData.cs
@@ -3,6 +3,6 @@
     public class ModBusData
     {
         public ushort Address{ get; set; }
-        public byte Value { get; set; }
+        public ushort Value { get; set; }
     }
 }
